+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme"
-            android:name=".CustomContext">
+            android:name=".contract.CustomContext">
         <activity android:name=".activity.ListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
