+++ app/src/main/java/com/example/vakulenko/notebook/dao/AppDatabase.java
@@ -2,7 +2,7 @@ package com.example.vakulenko.notebook.dao;
 
 import android.arch.persistence.room.Database;
 import android.arch.persistence.room.RoomDatabase;
-import com.example.vakulenko.notebook.entitys.Note;
+import com.example.vakulenko.notebook.dao.entitys.Note;
 
 @Database(entities = {Note.class}, version = 1)
 public abstract class AppDatabase extends RoomDatabase {
